LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerrawupdate-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
pointerrawupdate

Test Description: This test checks if pointerrawupdate is dispatched correctly.

Move your mouse within the black box.

Press left button down and then press middle button while holding down left button. Then release the buttons


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointerrawupdate event received
